+++ src/app/home/Home.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { SideBar } from '@shared/ui/sidebar/SideBar.tsx';
 import { Header } from '@shared/ui/header/Header.tsx';
 import { Footer } from '@shared/ui/footer/Footer.tsx';
-import { Hero } from '@src/components/hero/Hero.tsx';
+import { Hero } from '@app/home/hero/Hero.tsx';
 import { MainContainer } from '@shared/layout/MainContainer.tsx';
 
 interface HomeProps {
@@ -13,7 +13,9 @@ interface HomeProps {
 export const Home = ({ appearance, setAppearance }: HomeProps) => {
   return (
     <MainContainer
-      sidebar={<SideBar appearance={appearance} setAppearance={setAppearance} />}
+      sidebar={
+        <SideBar appearance={appearance} setAppearance={setAppearance} />
+      }
     >
       <Header />
       <Hero />
