+++ src/domain/user/components/signup/Signup.tsx
@@ -1,0 +1,3 @@
+export const Signup = () => {
+  return <>Signup</>;
+};
