+++ src/domain/user/pages/signup/SignupPage.tsx
@@ -1,0 +1,3 @@
+export const SignupPage = () => {
+  return <>signup page</>;
+};
